requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aiosmtplib==3.0.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.4.0
arxiv==1.4.8
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cryptography==36.0.0
cycler==0.12.1
cymem==2.0.8
dashscope==1.19.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
edge-tts==6.1.12
email_validator==2.2.0
fastapi==0.110.0
feedparser==6.0.11
ffmpy==0.3.2
filelock==3.14.0
FLAML==2.1.2
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gradio @ https://public.agent-matrix.com/publish/gradio-3.32.10-py3-none-any.whl#sha256=960693cb59afc9f5e1c4aa7c3ff40857148bdfcdbbe10a9e663e56ab174b9a36
gradio_client==0.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.4.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
latex2mathml==3.77.0
linkify-it-py==2.0.3
lxml==5.2.2
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdit-py-plugins==0.3.3
mdtex2html==1.3.0
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.10
numpy==1.26.4
openai==1.31.0
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
preshed==3.0.9
prompt_toolkit==3.0.46
protobuf==3.18.0
pyautogen==0.2.28
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pymdown-extensions==10.8.1
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyPDF2==2.12.1
pyphen==0.15.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-markdown-math==0.8
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rjsmin==1.2.2
rpds-py==0.18.1
safetensors==0.4.3
scipdf-parser==0.52
semantic-version==2.10.0
setuptools==70.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.36.3
termcolor==2.4.0
textstat==0.7.3
thinc==8.2.4
tiktoken==0.7.0
tokenizers==0.19.1
toolz==0.12.1
tqdm==4.66.4
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.30.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.4
zhipuai==2.0.1
